requirements.txt
torch
torchvision
tqdm
matplotlib
scikit-learn
grad-cam
pillow
streamlit